update on pr comments
