+++ src/main/resources/reference.conf
@@ -7,16 +7,6 @@ ensembl_version = "YYY"
 efo_version = "MM-mm-pp"
 // --- END --- Release baseline parameters --- //
 
-// Bucket name information
-bucketName = "open-targets-pre-data-releases"
-bucketPathBase = "gs://"${bucketName}"/"${releaseNumber}
-pathInputsLiterature = ${bucketPathBase}"/input/literature-inputs"
-pathBasePlatformOutput = ${bucketPathBase}"/output/etl/parquet"
-pathBaseOutputLiterature = ${bucketPathBase}"/output/literature-etl/"
-
-// Release information
-releaseNumber = "22.04"
-epmcReleaseNumber = "22.01"
 
 
 spark-settings {
@@ -1300,6 +1290,17 @@ otarproject {
 }
 
 literature {
+
+  // Bucket name information
+  bucketName = "open-targets-pre-data-releases"
+  bucketPathBase = "gs://"${literature.bucketName}"/"${literature.releaseNumber}
+  pathInputsLiterature = ${literature.bucketPathBase}"/input/literature-inputs"
+  pathBasePlatformOutput = ${literature.bucketPathBase}"/output/etl/parquet"
+
+  // Release information
+  releaseNumber = "22.04"
+  epmcReleaseNumber = "22.01"
+
   common = {
     publication-section-ranks = [
       {section: "title", rank: 1, weight: 1}
@@ -1319,7 +1320,7 @@ literature {
   processing {
     epmcids {
       format = "csv"
-      path = ${pathInputsLiterature}"/PMID_PMCID_DOI.csv.gz"
+      path = ${literature.pathInputsLiterature}"/PMID_PMCID_DOI.csv.gz"
       options = [
         {k: "header", v: "true"}
         {k: "inferSchema", v: "true"}
@@ -1327,20 +1328,20 @@ literature {
     }
     diseases {
       format = "parquet"
-      path = ${pathBasePlatformOutput}"/diseases"
+      path = ${literature.pathBasePlatformOutput}"/diseases"
     }
 
     targets {
       format = "parquet"
-      path = ${pathBasePlatformOutput}"/targets"
+      path = ${literature.pathBasePlatformOutput}"/targets"
     }
     drugs {
       format = "parquet"
-      path = ${pathBasePlatformOutput}"/molecule"
+      path = ${literature.pathBasePlatformOutput}"/molecule"
     }
     epmc {
       format = "json"
-      path = "gs://otar025-epmc/"${epmcReleaseNumber}"/**/"
+      path = "gs://otar025-epmc/"${literature.epmcReleaseNumber}"/**/"
     }
     outputs = {
       raw-evidence {
