+++ src/test/scala/io/opentargets/etl/backend/literature/LiteratureTest.scala
@@ -2,6 +2,7 @@ package io.opentargets.etl.backend.literature
 
 import io.opentargets.etl.backend.EtlSparkUnitTest
 import org.scalatest.matchers.should.Matchers._
+import org.scalatest.Inspectors._
 
 class LiteratureTest extends EtlSparkUnitTest {
 
@@ -13,12 +14,12 @@ class LiteratureTest extends EtlSparkUnitTest {
     session should not be (null)
 
     val sessionConfigurations = session.sparkSession.conf.getAll
-    val configurations = session.configuration.literature.common.sparkSessionConfig
-      .getOrElse(Seq())
+    val configurations = session.configuration.literature.common.sparkSessionConfig.get
       .map((conf) => (conf.k, conf.v))
       .toMap
+      .keys
 
-    configurations.foreach(conf => sessionConfigurations should contain(conf))
+    forAll(configurations)(conf => sessionConfigurations should contain key (conf))
 
   }
